+++ geotrek/signage/tests/test_command_signage.py
@@ -16,7 +16,7 @@ from geotrek.common.tests.utils import update_gis
 
 class SignageCommandTest(TestCase):
     """
-    There are 2 signages in the file signage.shp
+    There are 3 signages in the file signage.shp
     """
     @classmethod
     def setUpTestData(cls):
@@ -27,24 +27,27 @@ class SignageCommandTest(TestCase):
         structure = StructureFactory.create(name='structure')
         filename = os.path.join(os.path.dirname(__file__), 'data', 'signage.shp')
         call_command('loadsignage', filename, type_default='label', name_default='name',
-                     condition_default='condition', structure_default='structure',
+                     condition_default='conditions', condition_field='conditions', structure_default='structure',
                      description_default='description', year_default=2010, verbosity=2, stdout=output)
         self.assertIn('Signages will be linked to %s' % structure, output.getvalue())
-        self.assertIn('2 objects created.', output.getvalue())
-        value = Signage.objects.filter(name='name')
-        self.assertEqual(2010, value[0].implantation_year)
-        self.assertEqual(value.count(), 2)
-        self.assertAlmostEqual(value[0].geom.x, -436345.704831, places=5)
-        self.assertAlmostEqual(value[0].geom.y, 1176487.742917, places=5)
-        self.assertAlmostEqual(value[1].geom.x, -436345.505347, places=5)
-        self.assertAlmostEqual(value[1].geom.y, 1176480.918334, places=5)
+        self.assertIn('3 objects created.', output.getvalue())
+        qs_signages = Signage.objects.filter(name='name')
+        self.assertEqual(2010, qs_signages[0].implantation_year)
+        self.assertEqual(qs_signages.count(), 3)
+        self.assertAlmostEqual(qs_signages[0].geom.x, -436345.704831, places=5)
+        self.assertAlmostEqual(qs_signages[0].geom.y, 1176487.742917, places=5)
+        self.assertAlmostEqual(qs_signages[1].geom.x, -436345.505347, places=5)
+        self.assertAlmostEqual(qs_signages[1].geom.y, 1176480.918334, places=5)
+        self.assertEqual("condition", qs_signages[0].conditions.first().label)
+        self.assertEqual("condition", qs_signages[1].conditions.first().label)
+        self.assertEqual("condition", qs_signages[2].conditions.first().label)
 
     def test_load_signage_multipoints(self):
         output = StringIO()
         structure = StructureFactory.create(name='structure')
         filename = os.path.join(os.path.dirname(__file__), 'data', 'signage_good_multipoint.geojson')
         call_command('loadsignage', filename, type_default='label', name_default='name',
-                     condition_default='condition', structure_default='structure',
+                     condition_default='conditions', structure_default='structure',
                      description_default='description', year_default=2010, verbosity=2, stdout=output)
         self.assertIn('Signages will be linked to %s' % structure, output.getvalue())
         self.assertIn('1 objects created.', output.getvalue())
@@ -69,7 +72,7 @@ class SignageCommandTest(TestCase):
                 output_file_path,
                 type_default='label',
                 name_default='name',
-                condition_default='condition',
+                condition_default='conditions',
                 structure_default='structure',
                 description_default='description',
                 year_field='year',
@@ -97,7 +100,7 @@ class SignageCommandTest(TestCase):
                     output_file_path,
                     type_default='label',
                     name_default='name',
-                    condition_default='condition',
+                    condition_default='conditions',
                     structure_default='structure',
                     description_default='description',
                     year_field='year',
@@ -110,7 +113,7 @@ class SignageCommandTest(TestCase):
         filename = os.path.join(os.path.dirname(__file__), 'data', 'signage_bad_multipoint.geojson')
         with self.assertRaisesRegex(CommandError, 'One of your geometry is a MultiPoint object with multiple points'):
             call_command('loadsignage', filename, type_default='label', name_default='name',
-                         condition_default='condition', structure_default='structure',
+                         condition_default='conditions', structure_default='structure',
                          description_default='description', year_default=2010, verbosity=2, stdout=output)
 
     def test_load_signage_with_fields(self):
@@ -118,8 +121,9 @@ class SignageCommandTest(TestCase):
         structure = StructureFactory.create(name='structure')
         filename = os.path.join(os.path.dirname(__file__), 'data', 'signage.shp')
         call_command('loadsignage', filename, type_field='label', name_field='name',
-                     condition_field='condition', manager_field='manager', sealing_field='sealing',
                      structure_default='structure',
+                     eid_field='eid',
+                     condition_field='conditions', manager_field='manager', sealing_field='sealing',
                      description_field='descriptio', year_field='year', code_field='name', verbosity=1, stdout=output)
         self.assertIn('Signages will be linked to %s' % structure, output.getvalue())
         self.assertIn("SignageType 'type' created", output.getvalue())
@@ -206,7 +210,7 @@ class SignageCommandTest(TestCase):
         call_command('loadsignage', filename, eid_field='eid', type_default='label',
                      name_default='name', verbosity=2, stdout=output)
         self.assertIn("Update : name with eid eid1", output.getvalue())
-        self.assertEqual(Signage.objects.count(), 2)
+        self.assertEqual(Signage.objects.count(), 3)
 
     def test_fail_structure_default_do_not_exist(self):
         output = StringIO()